third_party/WebKit/LayoutTests/editing/selection/selection-after-split-text-node-expected.txt

PASS selection.type is "Range"
PASS selection.anchorNode is newText
PASS selection.anchorOffset is 3
PASS selection.focusNode is newText
PASS selection.focusOffset is 6
PASS selection.toString() is "baz"
PASS successfullyParsed is true

TEST COMPLETE

PASS successfullyParsed is true

TEST COMPLETE
